Navigation entre lista de recetas y detalle. Uso de BD para levantar datos

+++ app/src/main/java/ar/edu/mdp/fi/celiaquia/modelo/Receta.kt
@@ -6,7 +6,7 @@ import java.io.Serializable
 
 @Entity
 data class Receta (
-    @PrimaryKey val uid: Int,
+    @PrimaryKey val uid: Long,
     val imagen: Int,
     val nombre: String,
     val descripcion: String,
